+++ NEWS.md
@@ -1,5 +1,10 @@
 
+
+# surveybootstrap 0.0.2
+
 * moved unit tests for goc and helper functions in from networkreporting
+* added rescaled.bootstrap.weights helper function for case when a tibble with all of the rescaled bootstrap weights is useful
+* added dependencies to tidyverse, magrittr, and purrr
 
 # surveybootstrap 0.0.1